requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
bcrypt==4.3.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
email_validator==2.2.0
fastapi==0.119.0
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2==2.9.11
pwdlib==0.2.1
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing_extensions==4.15.0
uvicorn==0.37.0
# Vintern-1B-v3.5 model dependencies
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.20.0
einops>=0.6.0
timm>=0.9.0
# Image processing dependencies (for Vintern preprocessing)
numpy<2.0.0,>=1.21.0
pillow>=10.2.0
# Gemini AI for classification and explanation
google-genai>=0.2.0
# Web scraping for whitelist crawling
requests>=2.32.0
beautifulsoup4>=4.12.0
